+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [[0.8.1]] - 2021-05-30
 ### Fixed
 - Now `ImageFullView` will not catch mouse input in the huge box area around it
+- Now file input changes inner text to the path of selected file for upload.
 
 ## [[0.8.0]] - 2021-05-28
 ### Added